requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.20
botocore==1.40.20
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
jmespath==1.0.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
